pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sh40-celx"
description = "A TUI framework built for hypermedia."
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "bczsalba", email = "bczsalba@gmail.com" },
]
classifiers = [
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS",
  "Operating System :: POSIX :: Linux",
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
  "Typing :: Typed",
  "Topic :: Software Development :: Libraries",
  "Topic :: Terminals",
  "Topic :: Text Processing :: Markup",
]
dependencies = ["sh40-celadon", "requests"]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/shade40/celx#readme"
Issues = "https://github.com/shade40/celx/issues"
Source = "https://github.com/shade40/celx"

[tool.hatch.version]
path = "celx/__about__.py"

[tool.hatch.build]
include = [
  "celx/*.py",
  "celx/py.typed",
  "/tests",
]

[tool.hatch.envs.default]
dependencies = [
  "mypy",
  "pylint",
  "pytest",
  "pytest-cov",
]

[tool.hatch.envs.test]
dependencies = [
  "mypy",
  "pylint",
  "pytest",
  "pytest-cov",
]

[tool.hatch.envs.default.scripts]
test = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=celx --cov=tests && coverage html"
lint = "pylint celx"
type = "mypy celx"
upload = "hatch build && twine upload dist/* && hatch clean"

[[tool.hatch.envs.test.matrix]]
python = ["38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "celx/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "def __repr__",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
